Just filter out config files not starting with sg_ and not ending with .yml
Closes #49

This is a refinement for the impl for #34. 

The original implementation had an own list of config types and checked based on these whether a file is "known" or not. However, it thus bypassed a more flexible implementation which also used the backend to verify the config type.

We now again use this implementation; the additional implementation just checks that the file name starts with sg_ and ends with .yml.

See merge request search-guard/sgctl!79